+++ deployments/terraform/aws/modules/cloudhsm/outputs.tf
@@ -1,0 +1,7 @@
+output "cluster_id" {
+  value = aws_cloudhsm_v2_cluster.this.cluster_id
+}
+
+output "security_group_id" {
+  value = aws_security_group.this.id
+}